% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAllCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create the exposure and outcome cohorts}
\usage{
createCohorts(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  cohortTable = "cohort", oracleTempSchema, outputFolder)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create the exposure and outcome cohorts
}
\details{
This function will create the exposure and outcome cohorts following the definitions included in
this package.
}
