% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{exportResults}
\alias{exportResults}
\title{Export all results to tables}
\usage{
exportResults(outputFolder, databaseId, databaseName, databaseDescription,
  minCellCount = 5, maxCores)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}

\item{maxCores}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}
}
\description{
Outputs all results to a folder called 'export', and zips them.
}
