% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{generateDiagnostics}
\alias{generateDiagnostics}
\title{Generate diagnostics}
\usage{
generateDiagnostics(outputFolder, maxCores)
}
\arguments{
\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{maxCores}{How many parallel cores should be used? If more cores are made
available this can speed up the analyses.}
}
\description{
Generate diagnostics
}
\details{
This function generates analyses diagnostics. Requires the study to be executed first.
}
