% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubmitResults.R
\name{submitResults}
\alias{submitResults}
\title{Submit the study results to the study coordinating center}
\usage{
submitResults(outputFolder, key, secret)
}
\arguments{
\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{key}{The key string as provided by the study coordinator}

\item{secret}{The secret string as provided by the study coordinator}
}
\value{
TRUE if the upload was successful.
}
\description{
Submit the study results to the study coordinating center
}
\details{
This will upload the file \code{StudyResults.zip} to the study coordinating center using Amazon S3.
This requires an active internet connection.
}
